+++ src/Network/Network/LoadBalancer/NewAzureRMLoadBalancerCommand.cs
@@ -95,6 +95,16 @@ namespace Microsoft.Azure.Commands.Network
 
         [Parameter(
             Mandatory = false,
+            HelpMessage = "The scope of the load balancer: external ('Public') or internal ('Private'). Set together with '-Mode Advanced' to create an advanced (Banksy-based) load balancer.",
+            ValueFromPipelineByPropertyName = true)]
+        [PSArgumentCompleter(
+            "Public",
+            "Private"
+        )]
+        public string Scope { get; set; }
+
+        [Parameter(
+            Mandatory = false,
             HelpMessage = "Object representing the frontend IPs to be used for the load balancer",
             ValueFromPipelineByPropertyName = true)]
         public PSFrontendIPConfiguration[] FrontendIpConfiguration { get; set; }
@@ -249,6 +259,7 @@ namespace Microsoft.Azure.Commands.Network
                 OutboundRules = this.OutboundRule?.ToList(),
                 Sku = vSku,
                 Mode = this.Mode,
+                Scope = this.Scope,
                 ExtendedLocation = string.IsNullOrEmpty(this.EdgeZone) ? null : new PSExtendedLocation(this.EdgeZone)
             };