+++ src/Network/Network/Models/PSLoadBalancer.cs
@@ -27,6 +27,8 @@ namespace Microsoft.Azure.Commands.Network.Models
         public PSLoadBalancerSku Sku { get; set; }
         [Ps1Xml(Target = ViewControl.Table)]
         public string Mode { get; set; }
+        [Ps1Xml(Target = ViewControl.Table)]
+        public string Scope { get; set; }
         public List<PSFrontendIPConfiguration> FrontendIpConfigurations { get; set; }
         public List<PSBackendAddressPool> BackendAddressPools { get; set; }
         public List<PSLoadBalancingRule> LoadBalancingRules { get; set; }